+++ BBHTTP/BBHTTPResponse.h
@@ -47,7 +47,7 @@ BBHTTPProtocolVersion BBHTTPProtocolVersionFromNSString(NSString* string);
 @property(strong, nonatomic, readonly) NSString* message;
 @property(strong, nonatomic, readonly) NSDictionary* headers;
 @property(assign, nonatomic, readonly, getter = isSuccessful) BOOL successful;
-@property(assign, nonatomic, readonly) NSUInteger contentSize;
+@property(assign, nonatomic, readonly) unsigned long long contentSize;
 @property(strong, nonatomic, readonly) id content;
 
 
@@ -65,7 +65,7 @@ BBHTTPProtocolVersion BBHTTPProtocolVersionFromNSString(NSString* string);
 
 #pragma mark Interface
 
-- (void)finishWithContent:(id)content size:(NSUInteger)size successful:(BOOL)successful;
+- (void)finishWithContent:(id)content size:(unsigned long long)size successful:(BOOL)successful;
 - (NSString*)headerWithName:(NSString*)header;
 - (void)setValue:(NSString*)value forHeader:(NSString*)header;
 - (NSString*)objectForKeyedSubscript:(NSString*)header;
